Make the project VS6 compatible
This enables the project to be compiled on Windows XP with VS6 and to
target period correct toolchain.

The project will not compile using VS6 right away, because it will need
some typedefs force included, but that's trivial to do via CMake.

The resulting .dll will import msvcrt.dll and be a bit bigger than it is
right now, which is 44 KBs with /O1 and /G6.