Replace stb_ds.h with gperf generated lookups
The gperf version used here comes from
https://github.com/rurban/gperf/tree/intkeys and requires some patches
to prevent segfaults, however even then the output needs to be modified
by hand a little for int keys.

Because this fork doesn't work properly for keys like 0x80000040, the
keys are indices with a base and there are helper macros to properly
manage that.